Updated Dockerfile, main.py, and bot version
Updated the Dockerfile to change the version label from 1.6.1 to 1.6.2. In main.py, the datetime module has been added to the import statements and removed from a later part. The bot version has been updated from 1.6.1 to 1.6.2 in the sentry_sdk.init() function. The start_time variable in the aclient class now uses the datetime module correctly. The URL for the GitLab repository in the self function has been updated.